% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read.R
\docType{data}
\name{ageclassMap}
\alias{ageclassMap}
\title{Mapping Age Classes data.frame
\code{ageclass}: original age classes from BAG data sources
\code{AgeClass}: re-categorized Age Classes for the 1st part of the report
\code{AgeClass2}: re-categorized Age Classes for the 2nd part of the report}
\format{
An object of class \code{data.frame} with 9 rows and 3 columns.
}
\usage{
ageclassMap
}
\description{
Mapping Age Classes data.frame
\code{ageclass}: original age classes from BAG data sources
\code{AgeClass}: re-categorized Age Classes for the 1st part of the report
\code{AgeClass2}: re-categorized Age Classes for the 2nd part of the report
}
\keyword{datasets}
