% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_to_month}
\alias{aggregate_to_month}
\title{Compute Totals over a given time period (month)}
\usage{
aggregate_to_month(
  data,
  pd,
  by = c("AgeClass", "vaccination_status"),
  aggv,
  cumv
)
}
\arguments{
\item{data}{data.frame data}

\item{pd}{character period label to mutate in Week column}

\item{by}{character vector of column names to group_by}

\item{aggv}{character vector of columns that can be aggregated (daily record)}

\item{cumv}{character vector of columns that are cumulative (daily record) and should not be aggregated}
}
\description{
Compute Totals over a given time period (month)
}
