% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read.R
\docType{data}
\name{bag_api_url}
\alias{bag_api_url}
\title{URL BAG website where data are available}
\format{
An object of class \code{character} of length 1.
}
\usage{
bag_api_url
}
\description{
URL BAG website where data are available
}
\keyword{datasets}
