% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{calc_ratio_fullyvac}
\alias{calc_ratio_fullyvac}
\title{Calculates ratio of hosp and death between Unvaccinated and Vaccinated}
\usage{
calc_ratio_fullyvac(
  data,
  ageclassmap = ageclassMap,
  refstatus = "not_vaccinated"
)
}
\arguments{
\item{data}{data.frame data}

\item{ageclassmap}{data.frame Age class mapping, used for factor levels}

\item{refstatus}{character vector of reference vaccination status for the computation}
}
\description{
Calculates ratio of hosp and death between Unvaccinated and Vaccinated
}
