% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{color_cells}
\alias{color_cells}
\title{Color table cells}
\usage{
color_cells(
  data,
  header,
  cgroup = NULL,
  rnames,
  rgbn = 255,
  skip = "All",
  table_cell_css = ""
)
}
\arguments{
\item{data}{data.frame data}

\item{header}{character, variable name for header}

\item{cgroup}{character, variable name for cgroup if present, NULL}

\item{rnames}{character, variable level to skip}

\item{rgbn}{numeric RGBN value}

\item{table_cell_css}{character feature of css cells}
}
\description{
Color table cells
}
