% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.funformat}
\alias{.funformat}
\title{Formats numbers with thousands separator or with \%}
\usage{
.funformat(x, perc, digits = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{perc}{logical if TRUE then \%}

\item{digits}{integer, if perc == TRUE then controls the rounding of the figure}
}
\description{
Formats numbers with thousands separator or with \%
}
