% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_lineplot.R
\name{filter_dataplot}
\alias{filter_dataplot}
\title{Filter line plot data function}
\usage{
filter_dataplot(dat, status, case)
}
\arguments{
\item{dat}{data.frame input data.}

\item{status}{character vaccination status as choice in selectInput.}

\item{case}{character Hospitalization or Deaths case as choice in selectInput.}
}
\description{
Filter line plot data function
}
