% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{includeRMarkdown}
\alias{includeRMarkdown}
\title{Load rendered RMarkdown from a file and turn into HTML.}
\usage{
includeRMarkdown(path)
}
\arguments{
\item{path}{character path of Rmd}
}
\description{
Load rendered RMarkdown from a file and turn into HTML.
}
