% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{LinePlotCovid}
\alias{LinePlotCovid}
\title{Line Plot function}
\usage{
LinePlotCovid(
  df,
  FACET = "AgeClass",
  g_palette,
  percent = FALSE,
  title = "Time-line of Records per Age class"
)
}
\arguments{
\item{df}{data.frame data}

\item{FACET}{character, variable name for the facets}

\item{g_palette}{character vector of colors for the graph}

\item{percent}{logical if TRUE then data and labels are in \%}

\item{title}{character}
}
\description{
Line Plot function
}
