% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{make_100k}
\alias{make_100k}
\title{Re-scales values to 100'000 people}
\usage{
make_100k(data, by = c("AsOfDate", "AgeClass", "vaccination_status"), status)
}
\arguments{
\item{data}{data.frame data}

\item{by}{character vector of column names to group_by}

\item{status}{character vector of Vaccination Status column name}
}
\description{
Re-scales values to 100'000 people
}
