% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ProjCol}
\alias{ProjCol}
\title{Color for Current Status}
\format{
An object of class \code{character} of length 1.
}
\usage{
ProjCol
}
\description{
Color for Current Status
}
\keyword{datasets}
