% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read.R
\name{read_bag_data_cases}
\alias{read_bag_data_cases}
\title{Read Infection cases data from BAG source}
\usage{
read_bag_data_cases(bag.admin.url, ageclassMap, dateweek = NULL)
}
\arguments{
\item{bag.admin.url}{character vector, BAG data url}

\item{ageclassMap}{data.frame with mapping age}

\item{dateweek}{dates in Weeks format to be used to filter the data}
}
\description{
Read Infection cases data from BAG source
}
