% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read.R
\name{read_bag_data_vac}
\alias{read_bag_data_vac}
\title{Read Vaccination data from BAG source}
\usage{
read_bag_data_vac(bag.admin.url, ageclassMap)
}
\arguments{
\item{bag.admin.url}{character vector, BAG data url}

\item{ageclassMap}{data.frame with mapping age}
}
\description{
Read Vaccination data from BAG source
}
