% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_unkwnown}
\alias{rescale_unkwnown}
\title{Allocates proportionally values of Unknown vaccination status to the other Vaccination Categories}
\usage{
rescale_unkwnown(data, by = c("AsOfDate", "AgeClass", "Case"))
}
\arguments{
\item{data}{data.frame data}

\item{by}{character vector of column names to group_by}
}
\description{
Allocates proportionally values of Unknown vaccination status to the other Vaccination Categories
}
