% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{It renders the Article Rmd report}
\usage{
run_report(rmdname = "Index.Rmd", quiet = TRUE, run_pandoc = TRUE)
}
\arguments{
\item{rmdname}{character name of the report}

\item{quiet}{logical suppress printing}

\item{run_pandoc}{logical run_pandoc argument of render}
}
\description{
It renders the Article Rmd report
}
