% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{StackedBarplotCovid}
\alias{StackedBarplotCovid}
\title{Stacked Bar plot function}
\usage{
StackedBarplotCovid(
  df,
  X,
  FILL,
  FACET,
  percent = FALSE,
  g_palette,
  position = "fill",
  title = ""
)
}
\arguments{
\item{df}{data.frame data}

\item{X}{character, variable name in X axis}

\item{FILL}{character, variable name for position fill}

\item{FACET}{character, variable name for the facets}

\item{percent}{logical if TRUE then data and labels are in \%}

\item{g_palette}{character vector of colors for the graph}

\item{position}{character position of bar plot, fill}

\item{title}{character}
}
\description{
Stacked Bar plot function
}
