% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vac_levels}
\alias{vac_levels}
\alias{vac_levels_colors}
\alias{pick_vac_lev}
\alias{pick_vac_lev_col}
\title{Vaccination status vector with labels as names}
\usage{
vac_levels()

vac_levels_colors()

pick_vac_lev(x)

pick_vac_lev_col(x)
}
\arguments{
\item{x}{character vaccination status from \code{vac_levels()}}
}
\description{
Vaccination status vector with labels as names

Vaccination status color named vector

Pick Label given Vaccination status

Pick color given vaccination status
}
