% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weeks_to_date}
\alias{weeks_to_date}
\title{Utility that converts weeks in form YY-W-\link{0-9} into real dates in form yy/mm/dd}
\usage{
weeks_to_date(w, range = TRUE)
}
\arguments{
\item{w}{date vector}

\item{range}{logical if TRUE then only 1st and last values are returned}
}
\description{
Utility that converts weeks in form YY-W-\link{0-9} into real dates in form yy/mm/dd
}
