% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data.R
\name{build_data}
\alias{build_data}
\title{It renders the Article Rmd report}
\usage{
build_data(url = bag_api_url, agemap = ageclassMap, writerds = TRUE)
}
\arguments{
\item{url}{character bag_api_url}

\item{agemap}{data.frame ageclassMap Age mapping}

\item{writerds}{logical write RDS files}
}
\description{
It renders the Article Rmd report
}
