% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Run the Rmd report}
\usage{
run_report(...)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:run]{rmarkdown::run()}}.}
}
\description{
Run the Rmd report
}
