% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Run the Rmd report}
\usage{
run_report(
  fetch_latest_data = FALSE,
  render_args = NULL,
  auto_reload = TRUE,
  shiny_args = NULL
)
}
\arguments{
\item{fetch_latest_data}{logical, if FALSE the data are read from source, if TRUE from RDS files in \code{bag_data} folder. Default = TRUE, FALSE to be used in local run.}

\item{render_args}{list, additional arguments to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}}

\item{shiny_args}{list, additional arguments to pass to \code{\link[shiny:runApp]{shiny::runApp()}}}
}
\description{
Run the Rmd report
}
